refactor: Extract field configs to separate files
