+++ packages/form-builder/src/blocks/fields/email.js
@@ -1,0 +1,30 @@
+import {Icon} from '@wordpress/icons';
+import {__} from "@wordpress/i18n";
+import settings from "./settings";
+
+const email = {
+    name: 'custom-block-editor/email-field',
+    category: 'input',
+    settings: {
+        ...settings,
+        title: __('Email', 'custom-block-editor'),
+        supports: {
+            multiple: false,
+        },
+        attributes: {
+            lock: {remove: true},
+            label: {
+                default: __('Email Address'),
+            },
+        },
+        icon: () => <Icon icon={
+            <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path
+                    d="M22 6C22 4.9 21.1 4 20 4H4C2.9 4 2 4.9 2 6V18C2 19.1 2.9 20 4 20H20C21.1 20 22 19.1 22 18V6ZM20 6L12 11L4 6H20ZM20 18H4V8L12 13L20 8V18Z"
+                    fill="#000C00" />
+            </svg>
+        } />,
+    },
+};
+
+export default email;
